requirements.txt
flask
flask-cors
langchain
langchain-community
langchain-openai
langchain-anthropic
python-dotenv
pydantic
duckduckgo-search
openai
anthropic
